Initial check-in of Tornados and Trailer Parks
It is a well known fact that when a tornado touches down the first thing it does is ask for 
directions to the nearest trailer park.  This is an unfortunate thing for people who live in trailer 
parks (as I did when I was a kid).  

There is a dataset out there that contains statistical information on every tornado in the US from 
1950 to 2015.  So it occured to me that this data might be useful in determining the best place to 
build a trailer park in the US.

Now I have come to understand from reputable people that tornados are spawned as a result of pre-
existing conditions.  They don't actually get to choose where they are going to touchdown.  So they
can't in a targeted way just drop in on a specific trailer park.

Therefore, there must exist some optimal location to build a trailer park which is removed from the
areas where tornados frequently spawn.

Our goal in this effort is to find that sweet spot if you will.